+++ src/main/java/com/inventrik/digitalestore/service/theme/StoreThemeService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface StoreThemeService {
     List<StoreThemeResponse> getAllThemes(Integer tenantId);
-    StoreThemeResponse getTheme(Integer tenantId, Integer themeId);
+    StoreThemeResponse getTheme(Integer tenantId, Long themeId);
     StoreThemeResponse createTheme(Integer tenantId, StoreThemeRequest request, String username);
-    StoreThemeResponse updateTheme(Integer tenantId, Integer themeId, StoreThemeRequest request, String username);
-    void deleteTheme(Integer tenantId, Integer themeId);
+    StoreThemeResponse updateTheme(Integer tenantId, Long themeId, StoreThemeRequest request, String username);
+    void deleteTheme(Integer tenantId, Long themeId);
 }
